+++ src/Presentation/Console/Command/InitCommand.php
@@ -16,7 +16,7 @@ class InitCommand extends AbstractCommand
 {
     private const DEFAULT_REL_VERSION = '2.x-dev';
 
-    protected function configure() : void
+    protected function configure()
     {
         $this->setName('bartlett:db:init')
             ->setDescription('Load JSON file(s) in SQLite database')
@@ -29,7 +29,7 @@ class InitCommand extends AbstractCommand
         ;
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output) : void
+    protected function execute(InputInterface $input, OutputInterface $output)
     {
         $initCommand = new AppInitCommand();
         $initCommand->extension = '';
@@ -45,5 +45,7 @@ class InitCommand extends AbstractCommand
         $initCommand->output = $output;
 
         $this->commandBus->handle($initCommand);
+
+        return 0;
     }
 }