+++ src/Presentation/Console/Command/ShowCommand.php
@@ -12,7 +12,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class ShowCommand extends AbstractCommand
 {
-    protected function configure() : void
+    protected function configure()
     {
         $this->setName('bartlett:db:show')
             ->setDescription('Show details of a reference supported in the Database.')
@@ -32,7 +32,7 @@ class ShowCommand extends AbstractCommand
         ;
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output) : void
+    protected function execute(InputInterface $input, OutputInterface $output)
     {
         $showCommand = new AppShowCommand(
             trim($input->getArgument('extension')),
@@ -91,7 +91,7 @@ class ShowCommand extends AbstractCommand
                 array($summary['static methods'])
             );
             $this->printFormattedLines($output, $summary);
-            return;
+            return 0;
         }
 
         foreach ($results as $title => $values) {
@@ -140,6 +140,8 @@ class ShowCommand extends AbstractCommand
             $this->tableHelper($output, $headers, $rows);
             $output->writeln('');
         }
+
+        return 0;
     }
 
     private static function ext(array $versions) : string