+++ CHANGELOG.md
@@ -6,10 +6,13 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 ## [Unreleased]
 
+### Changed
+
 - Amqp reference updated to version 1.10.0 (stable)
 - Xdebug reference updated to version 2.9.4 (stable)
 - Lzf reference updated to version 1.6.8 (stable)
 - Yaml reference updated to version 2.1.0 (stable)
+- remove Symfony 3.x compatibility and add support to Symfony 5.x
 
 ## [2.12.0] - 2020-03-21